pyproject.toml
[tool.poetry]
name = "aimmo_assignment"
version = "0.1.0"
description = ""
authors = ["Joowan Kim <rlawndhks217@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "3.2.9"
PyJWT = "2.3.0"
mongoengine = "0.23.1"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"
assertpy = "1.1"
mongomock = "3.23.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
